finish up listing creation and add listings profile page

+++ packages/subgraph/src/utils/eth.ts
@@ -1,9 +1,9 @@
-import { BigInt } from "@graphprotocol/graph-ts";
+import { BigDecimal, BigInt } from "@graphprotocol/graph-ts";
 import { ONE_ETH_IN_WEI } from "./constants";
 
 /**
  * Converts Wei to Ether
  */
-export const weiToEth = (amount: BigInt): BigInt => {
-  return amount.div(ONE_ETH_IN_WEI);
+export const weiToEth = (amount: BigInt): BigDecimal => {
+  return amount.toBigDecimal().div(ONE_ETH_IN_WEI.toBigDecimal());
 };
